move fastify utils into dir
