+++ src/bin/pohttp-server.ts
@@ -1,7 +1,7 @@
 // tslint:disable-next-line:no-var-requires
 require('make-promises-safe');
 
-import { runFastify } from '../services/fastifyUtils';
 import { makeServer } from '../services/pohttp/server';
+import { runFastify } from '../utilities/fastify';
 
 makeServer().then(runFastify);
