+++ src/services/poweb/server.spec.ts
@@ -1,5 +1,5 @@
 import { mockSpy } from '../../_test_utils';
-import * as fastifyUtils from '../fastifyUtils';
+import * as fastifyUtils from '../../utilities/fastify';
 import { setUpCommonFixtures } from './_test_utils';
 import RouteOptions from './RouteOptions';
 import { makeServer } from './server';
